Adding a code to read user-api-secret keys from config file. Used for external programs that use just api calls.
