+++ src/python/dart/web/api/auth.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, request, current_app, redirect, flash, url_for, render_template, g
 from flask.ext.login import LoginManager, login_user, logout_user
+from datetime import datetime
 
 from dart.service.user import UserService
 from dart.service.api_key import ApiKeyService
